+++ conn_test.go
@@ -300,6 +300,7 @@ func TestClose(t *testing.T) {
 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	require.NoError(t, conn.start(ctx))
 	cancel()
+	require.Nil(t, conn.Properties())
 	require.NoError(t, conn.Close())
 	// with Close error
 	netConn = fake.NewNetConn(senderFrameHandlerNoUnhandled(0, SenderSettleModeUnsettled), fake.NetConnOptions{})
@@ -1069,4 +1070,40 @@ func TestConnSmallFrames(t *testing.T) {
 	require.NoError(t, session.Close(ctx))
 	cancel()
 	require.NoError(t, conn.Close())
+}
+
+func TestConnProperties(t *testing.T) {
+	responder := func(remoteChannel uint16, req frames.FrameBody) (fake.Response, error) {
+		switch req.(type) {
+		case *fake.AMQPProto:
+			return newResponse(fake.ProtoHeader(fake.ProtoAMQP))
+		case *frames.PerformOpen:
+			b, err := fake.EncodeFrame(frames.TypeAMQP, 0, &frames.PerformOpen{
+				ChannelMax:   65535,
+				ContainerID:  "container",
+				IdleTimeout:  time.Minute,
+				MaxFrameSize: 4294967295,
+				Properties: map[encoding.Symbol]any{
+					"ConnProperty1": "foo",
+					"ConnProperty2": 123,
+				},
+			})
+			return newResponse(b, err)
+		case *frames.PerformClose:
+			return newResponse(fake.PerformClose(nil))
+		default:
+			return fake.Response{}, fmt.Errorf("unhandled frame %T", req)
+		}
+	}
+
+	netConn := fake.NewNetConn(responder, fake.NetConnOptions{})
+	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
+	client, err := NewConn(ctx, netConn, nil)
+	cancel()
+	require.NoError(t, err)
+	require.Equal(t, map[string]any{
+		"ConnProperty1": "foo",
+		"ConnProperty2": int64(123),
+	}, client.Properties())
+	require.NoError(t, client.Close())
 }